Draw multiple pages of test characters.

+++ pdfpage.hpp
@@ -83,7 +83,7 @@ public:
     void translate(double xtran, double ytran);
     void rotate(double angle);
     void render_utf8_text(std::string_view text, FontId fid, double pointsize, double x, double y);
-    void render_raw_glyph(uint16_t glyph, FontId fid, double pointsize, double x, double y);
+    void render_raw_glyph(uint32_t glyph, FontId fid, double pointsize, double x, double y);
     void render_ascii_text_builtin(
         const char *ascii_text, BuiltinFonts font_id, double pointsize, double x, double y);
